database-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cx_Oracle==8.3.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
pandas==2.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
python-dateutil==2.8.2
openpyxl==3.1.2
# LDAP Authentication
ldap3
requests 
# Stratification functionality
scikit-learn==1.3.2
scipy==1.11.4
numpy==1.24.4 
# Scheduling functionality
APScheduler==3.10.4
# Parquet file support
pyarrow==14.0.2 